--- src/components/ui/Select/City/SelectedPrefectureCity/index.tsx
@@ -1,18 +1,0 @@
-'use client';
-
-import { CityContent } from '@/components/ui/Select/City/Content';
-import { useGetCityListQuery } from './useGetCitiesQuery';
-import { FC } from 'react';
-
-//
-// TODO: 外部APIの都合上、自動反映を後回し（watchしているのでリストは作成される）
-//
-type Props = {
-  prefectureCode: string;
-};
-
-export const SelectedPrefectureCity: FC<Props> = ({ prefectureCode }) => {
-  const { data: response } = useGetCityListQuery(String(prefectureCode).padStart(2, '0'));
-
-  return <CityContent cityList={response?.data.data} />;
-};
